Make sure extensions are loaded automatically
Sublime only loads python packages from the top level of a Plugin
directory by default.

Dynamically importing packages after the plugin has loaded
is not possible. (See also
http://www.sublimetext.com/forum/viewtopic.php?f=6&t=6777 )

The solution is to add a package at the top level that automatically
loads packages from the extension directory of this plugin.